Add lodash-es dependency, update Carbon usage, and enhance various resources and services

+++ app/Http/Services/Master/MasterSearchService.php
@@ -80,7 +80,7 @@ class MasterSearchService
         return DB::select($query, $queryParams);
     }
 
-    private static function calculateSearchRadius(int $zoom): float
+    private static function calculateSearchRadius(float $zoom): float
     {
         $earthRadiusKm = 20037.5;
